+++ CHANGELOG.md
@@ -5,6 +5,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.11.0] - 2025-07-23
+
+### Added
+
+- Additional primitive type guards:
+  - `isBigint`
+  - `isNull` (same as `=== null`)
+  - `isObject` (same as `typeof value === "object"`)
+  - `isObjectNotNull`
+  - `isUndefined`
+
 ## [0.10.0] - 2025-07-02
 
 ### Added